Revert "Changed param back to int in moveForward"
This reverts commit ce270a8f43eaf3963b7952ab9bd809c3083c76e3.

+++ software/Mouse/Algo/API.c
@@ -51,7 +51,7 @@ int wallLeft() {
     return getBoolean("wallLeft");
 }
 
-int moveForward(int spaces) {
+int moveForward(double spaces) {
     printf("moveForward %d\n", spaces);
     fflush(stdout);
     char response[BUFFER_SIZE];
